Further clarifying env vars
