+++ core/chains/evm/config/ocr2_config.go
@@ -1,0 +1,10 @@
+package config
+
+func (c *chainScopedConfig) OCR2ContractConfirmations() uint16 {
+	val, ok := c.GeneralConfig.GlobalOCR2ContractConfirmations()
+	if ok {
+		c.logEnvOverrideOnce("OCR2ContractConfirmations", val)
+		return val
+	}
+	return c.defaultSet.ocr2ContractConfirmations
+}
