+++ core/cmd/local_client_test.go
@@ -135,8 +135,8 @@ LOG_SQL_MIGRATIONS: false
 LOG_SQL: false
 LOG_TO_DISK: true
 TRIGGER_FALLBACK_DB_POLL_INTERVAL: 30s
-OCR_CONTRACT_TRANSMITTER_TRANSMIT_TIMEOUT: 10s
-OCR_DATABASE_TIMEOUT: 10s
+OCR_CONTRACT_TRANSMITTER_TRANSMIT_TIMEOUT: 
+OCR_DATABASE_TIMEOUT: 
 OCR_DEFAULT_TRANSACTION_QUEUE_DEPTH: 1
 OCR_TRACE_LOGGING: false
 P2P_NETWORKING_STACK: V1
